+++ src/extensions/rcsb/graphql/types.ts
@@ -6,7 +6,7 @@ export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: 
 export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
 export type MakeEmpty<T extends { [key: string]: unknown }, K extends keyof T> = { [_ in K]?: never };
 export type Incremental<T> = T | { [P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never };
-// Generated on 2023-09-02T12:08:38-07:00
+// Generated on 2023-09-30T10:32:40-07:00
 
 /** All built-in and custom scalars, mapped to their actual values */
 export type Scalars = {