+++ src/mol-gl/webgl/context.ts
@@ -225,6 +225,8 @@ export interface WebGLContext {
     setContextLost: () => void
     handleContextRestored: (extraResets?: () => void) => void
 
+    setPixelScale: (value: number) => void
+
     /** Cache for compute renderables, managed by consumers */
     readonly namedComputeRenderables: { [name: string]: ComputeRenderable<any> }
     /** Cache for frambuffers, managed by consumers */
@@ -271,6 +273,8 @@ export function createContext(gl: GLRenderingContext, props: Partial<{ pixelScal
     let isContextLost = false;
     const contextRestored = new BehaviorSubject<now.Timestamp>(0 as now.Timestamp);
 
+    let pixelScale = props.pixelScale || 1;
+
     let readPixelsAsync: (x: number, y: number, width: number, height: number, buffer: Uint8Array) => Promise<void>;
     if (isWebGL2(gl)) {
         const pbo = gl.createBuffer();
@@ -315,7 +319,7 @@ export function createContext(gl: GLRenderingContext, props: Partial<{ pixelScal
         isWebGL2: isWebGL2(gl),
         get pixelRatio() {
             const dpr = (typeof window !== 'undefined') ? (window.devicePixelRatio || 1) : 1;
-            return dpr * (props.pixelScale || 1);
+            return dpr * (pixelScale || 1);
         },
 
         extensions,
@@ -355,6 +359,10 @@ export function createContext(gl: GLRenderingContext, props: Partial<{ pixelScal
 
             isContextLost = false;
             contextRestored.next(now());
+        },
+
+        setPixelScale: (value: number) => {
+            pixelScale = value;
         },
 
         createRenderTarget: (width: number, height: number, depth?: boolean, type?: 'uint8' | 'float32' | 'fp16', filter?: TextureFilter, format?: 'rgba' | 'alpha') => {